refactor: Extract retrieve_work patch
This should help DRY-up #175
